+++ Konata.Core/Konata.Core.csproj
@@ -40,6 +40,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Msf\Packets\Protobuf\ProtoGetMsg.cs" />
+    <Compile Include="Msf\Services\PbMessageSvc\PbMsgReadReport.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Bot.cs" />
     <Compile Include="AppInfo.cs" />
@@ -53,7 +54,7 @@
     <Compile Include="Msf\Services\MessageSvc\PbSendMsg.cs" />
     <Compile Include="Msf\Services\MessageSvc\PushForceOffline.cs" />
     <Compile Include="Msf\Services\MessageSvc\PushNotify.cs" />
-    <Compile Include="Msf\Services\MessageSvc\PushReaded.cs" />
+    <Compile Include="Msf\Services\MessageSvc\PushRead.cs" />
     <Compile Include="Msf\Services\MessageSvc\RequestPushStatus.cs" />
     <Compile Include="Msf\Services\OnlinePush\PbPushGroupMsg.cs" />
     <Compile Include="Msf\Services\PbMessageSvc\PbMessageWithDraw.cs" />
@@ -173,9 +174,9 @@
     <Compile Include="Msf\Packets\Oicq\OicqRequestCheckImage.cs" />
     <Compile Include="Msf\Packets\Oicq\OicqRequestCheckSms.cs" />
     <Compile Include="Msf\Packets\Oicq\OicqRequestTgtgt.cs" />
-    <Compile Include="Msf\Utils\Crypt\ICryptor.cs" />
-    <Compile Include="Msf\Utils\Crypt\Md5Cryptor.cs" />
-    <Compile Include="Msf\Utils\Crypt\TeaCryptor.cs" />
+    <Compile Include="Msf\Crypto\ICryptor.cs" />
+    <Compile Include="Msf\Crypto\Md5Cryptor.cs" />
+    <Compile Include="Msf\Crypto\TeaCryptor.cs" />
     <Compile Include="DeviceInfo.cs" />
     <Compile Include="Msf\UserSigInfo.cs" />
     <Compile Include="Utils\Guid.cs" />
