+++ Konata.Core/Msf/Packets/Oicq/OicqRequest.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using Konata.Library.IO;
-using Konata.Msf.Utils.Crypt;
+using Konata.Msf.Crypto;
 
 namespace Konata.Msf.Packets.Oicq
 {
