+++ Konata.Core/Msf/Packets/Oicq/OicqRequestTransport.cs
@@ -3,7 +3,6 @@ using System.IO;
 using System.IO.Compression;
 using Konata.Library.IO;
 using Konata.Msf.Crypto;
-using Konata.Msf.Utils.Crypt;
 
 namespace Konata.Msf.Packets.Oicq
 {
