+++ Konata.Core/Msf/Packets/Tlv/T202.cs
@@ -1,4 +1,4 @@
-﻿using Konata.Msf.Utils.Crypt;
+﻿using Konata.Msf.Crypto;
 
 namespace Konata.Msf.Packets.Tlv
 {
