+++ Konata.Core/Msf/Packets/Tlv/T545.cs
@@ -1,5 +1,5 @@
 ﻿using Konata.Utils;
-using Konata.Msf.Utils.Crypt;
+using Konata.Msf.Crypto;
 
 namespace Konata.Msf.Packets.Tlv
 {
