+++ Konata.Core/Msf/Packets/Tlv/Tlv.cs
@@ -1,5 +1,5 @@
 ﻿using Konata.Library.IO;
-using Konata.Msf.Utils.Crypt;
+using Konata.Msf.Crypto;
 
 namespace Konata.Msf.Packets.Tlv
 {
