+++ Konata.Core/Msf/Packets/ToServiceMessage.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using Konata.Msf.Utils.Crypt;
+using Konata.Msf.Crypto;
 
 namespace Konata.Msf.Packets
 {
